+++ xapian-core/backends/glass/glass_database.h
@@ -328,7 +328,7 @@ class GlassDatabase : public Xapian::Database::Internal {
 			int fd,
 			const std::vector<const Xapian::Database::Internal*>& sources,
 			const std::vector<Xapian::docid> & offset,
-			size_t block_size,
+			unsigned block_size,
 			Xapian::Compactor::compaction_level compaction,
 			unsigned flags,
 			Xapian::docid last_docid);
